third_party/WebKit/LayoutTests/http/tests/websocket/extensions.html

<!DOCTYPE html>
<html>
<head>
<script src="/js-test-resources/js-test.js"></script>
</head>
<body>
<div id="description"></div>
<div id="console"></div>
<script>
description("Test WebSocket.extensions attribute.");

window.jsTestIsAsync = true;

var ws = new WebSocket("ws://localhost:8880/echo-requested-extensions");
debug("Check the value before opening the connection:");
shouldBeEqualToString("ws.extensions", "");

// extensions attribute is read-only.
debug("Check if the value is read only:");
shouldBeEqualToString("ws.extensions = 'foo'; ws.extensions", "");

ws.onopen = function(event)
{
    debug("Check the value after the connection is established:");
    shouldBeEqualToString("ws.extensions", "permessage-deflate");
    ws.send("hi!");
};

ws.onmessage = function(event)
{
    debug("Check the value of the Sec-WebSocket-Extensions header sent to the server:");
    debug('event.data is "' + event.data + '"');
    finishJSTest();
};

</script>
</body>
</html>
